rewrite some attribites component in components/setting
