+++ mwdb/web/src/components/Settings/Views/AttributeEditTemplateView.tsx
@@ -1,19 +1,20 @@
-import React from "react";
 import { useNavigate, useOutletContext } from "react-router-dom";
 
 import { api } from "@mwdb-web/commons/api";
 import { useViewAlert } from "@mwdb-web/commons/hooks";
 
 import RichAttributePreview from "../../RichAttribute/RichAttributePreview";
+import { AttributeOutletContext } from "@mwdb-web/types/context";
 
-export function AttributeEditTemplate() {
-    const { attribute, getAttribute } = useOutletContext();
+export function AttributeEditTemplateView() {
+    const { attribute, getAttribute }: AttributeOutletContext =
+        useOutletContext();
     const { rich_template: richTemplate, example_value: exampleValue } =
         attribute;
     const navigate = useNavigate();
     const viewAlert = useViewAlert();
 
-    async function storeTemplate(template, value) {
+    async function storeTemplate(template: string, value: string) {
         try {
             await api.updateAttributeDefinition({
                 key: attribute.key,
